connection success with single user

+++ DecisionServer/QueueEntity.java
@@ -1,8 +1,8 @@
 public class QueueEntity
 {
 	int ID;
-	int var;
-	public QueueEntity(int ID_, int var_)
+	double var;
+	public QueueEntity(int ID_, double var_)
 	{
 		ID = ID_;
 		var = var_;
